Support translations with placeholders in PHP
The placeholder support was already added for JavaScript before but it is
required in PHP as well so it ports that code.

With the update to Google Maps JavaScript API version 3, the `$googleview` and
`$overview` variables have been changed. The former accepts different values
and the latter is deprecated. This is now issuing warnings about that.