Dynamically set total trip information
Instead of calculating the total distance in a separate loop it can determine
that from the added markers. It also counts the number of image or comment
markers in the JavaScript creator for the marker.

This simplifies the first query iteration and also makes the `distance`
function in the PHP script unnecessary.