Convert `findclosestpositionbytime` to PDO
This converts `findclosestpositionbytime` to use PDO and also simplifies the
SQL query. Previously it first got the dates which had the smallest difference
to the given date and then selected all rows based on those dates. But it
could just select all rows where the difference is the lowest difference
found so removing the middle step of just selecting the dates first.